feature-WebScraperStockSelector/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
mysqlclient==2.0.3
requests==2.26.0
beautifulsoup4==4.9.3
newspaper3k==0.2.8
python-dotenv==0.19.0
yfinance==0.1.63
transformers==4.11.3
torch==1.9.0
numpy==1.21.2
pandas==1.3.3
